+++ packages/destination-actions/src/destinations/amplitude/index.ts
@@ -7,64 +7,7 @@ import groupIdentifyUser from './groupIdentifyUser'
 import logPurchase from './logPurchase'
 import type { Settings } from './generated-types'
 import { getEndpointByRegion } from './common-functions'
-
 import logEventV2 from './logEventV2'
-
-/** used in the quick setup */
-const presets: DestinationDefinition['presets'] = [
-  {
-    name: 'Track Calls',
-    subscribe: 'type = "track" and event != "Order Completed"',
-    partnerAction: 'logEventV2',
-    mapping: defaultValues(logEventV2.fields),
-    type: 'automatic'
-  },
-  {
-    name: 'Order Completed Calls',
-    subscribe: 'type = "track" and event = "Order Completed"',
-    partnerAction: 'logPurchase',
-    mapping: defaultValues(logPurchase.fields),
-    type: 'automatic'
-  },
-  {
-    name: 'Page Calls',
-    subscribe: 'type = "page"',
-    partnerAction: 'logEventV2',
-    mapping: {
-      ...defaultValues(logEventV2.fields),
-      event_type: {
-        '@template': 'Viewed {{name}}'
-      }
-    },
-    type: 'automatic'
-  },
-  {
-    name: 'Screen Calls',
-    subscribe: 'type = "screen"',
-    partnerAction: 'logEventV2',
-    mapping: {
-      ...defaultValues(logEventV2.fields),
-      event_type: {
-        '@template': 'Viewed {{name}}'
-      }
-    },
-    type: 'automatic'
-  },
-  {
-    name: 'Identify Calls',
-    subscribe: 'type = "identify"',
-    partnerAction: 'identifyUser',
-    mapping: defaultValues(identifyUser.fields),
-    type: 'automatic'
-  },
-  {
-    name: 'Browser Session Tracking',
-    subscribe: 'type = "track" or type = "identify" or type = "group" or type = "page" or type = "alias"',
-    partnerAction: 'sessionId',
-    mapping: {},
-    type: 'automatic'
-  }
-]
 
 const destination: DestinationDefinition<Settings> = {
   name: 'Actions Amplitude',
@@ -125,7 +68,60 @@ const destination: DestinationDefinition<Settings> = {
       }
     })
   },
-  presets,
+  presets: [
+    {
+      name: 'Track Calls',
+      subscribe: 'type = "track" and event != "Order Completed"',
+      partnerAction: 'logEventV2',
+      mapping: defaultValues(logEventV2.fields),
+      type: 'automatic'
+    },
+    {
+      name: 'Order Completed Calls',
+      subscribe: 'type = "track" and event = "Order Completed"',
+      partnerAction: 'logPurchase',
+      mapping: defaultValues(logPurchase.fields),
+      type: 'automatic'
+    },
+    {
+      name: 'Page Calls',
+      subscribe: 'type = "page"',
+      partnerAction: 'logEventV2',
+      mapping: {
+        ...defaultValues(logEventV2.fields),
+        event_type: {
+          '@template': 'Viewed {{name}}'
+        }
+      },
+      type: 'automatic'
+    },
+    {
+      name: 'Screen Calls',
+      subscribe: 'type = "screen"',
+      partnerAction: 'logEventV2',
+      mapping: {
+        ...defaultValues(logEventV2.fields),
+        event_type: {
+          '@template': 'Viewed {{name}}'
+        }
+      },
+      type: 'automatic'
+    },
+    {
+      name: 'Identify Calls',
+      subscribe: 'type = "identify"',
+      partnerAction: 'identifyUser',
+      mapping: defaultValues(identifyUser.fields),
+      type: 'automatic'
+    },
+    {
+      name: 'Browser Session Tracking',
+      subscribe: 'type = "track" or type = "identify" or type = "group" or type = "page" or type = "alias"',
+      partnerAction: 'sessionId',
+      mapping: {},
+      type: 'automatic'
+    }
+  ],
   actions: {
     logEvent,
     identifyUser,
