+++ packages/destination-actions/src/destinations/amplitude/logEvent/index.ts
@@ -2,11 +2,11 @@ import type { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { send } from '../events-functions'
-import { common_fields } from '../common-fields'
-import { common_track_fields } from '../common-track-fields'
-import { common_track_identify_fields } from '../common-track-identify-fields'
-import { autocapture_fields } from '../autocapture-fields'
-import { min_id_length } from '../misc-fields'
+import { common_fields } from '../fields/common-fields'
+import { common_track_fields } from '../fields/common-track-fields'
+import { common_track_identify_fields } from '../fields/common-track-identify-fields'
+import { autocapture_fields } from '../fields/autocapture-fields'
+import { min_id_length } from '../fields/misc-fields'
 
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Log Event',
