+++ packages/destination-actions/src/destinations/amplitude/logPurchase/index.ts
@@ -2,12 +2,12 @@ import type { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { send } from '../events-functions'
-import { common_fields } from '../common-fields'
-import { common_track_fields } from '../common-track-fields'
-import { common_track_identify_fields } from '../common-track-identify-fields'
-import { autocapture_fields } from '../autocapture-fields'
+import { common_fields } from '../fields/common-fields'
+import { common_track_fields } from '../fields/common-track-fields'
+import { common_track_identify_fields } from '../fields/common-track-identify-fields'
+import { autocapture_fields } from '../fields/autocapture-fields'
 import { trackRevenuePerProduct } from './fields'
-import { min_id_length } from '../misc-fields'
+import { min_id_length } from '../fields/misc-fields'
 
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Log Purchase',
